ginkgo: rootdir: Override HWC service definition
CAF added a socket creation to hwcomposer service for post-processing
blobs to work. As AOSP interface does not have this, some display
features such as SVI configuration is broken and hwcomposer will crash
after adding it on boot. Instead of relying on framework fixes, we can
just override service definition and create the socket in this way.

Change-Id: I08ad6f808e566f9b35c932db473cd5dbf3af262a